Enable HA backup
The following code allows HA backup, so that we can test that we
can still backup in HA mode. Currently you can't restore in HA
mode without first pulling machines out of the replicaset and
then allowing restoring.

The commit tests pulling out of HA mode so that we have a tar file
backup to do so.